def cancel_shutdown

+++ exercAgendamento.py
@@ -17,11 +17,19 @@ def shutdown_computer():
 
 shutdown_computer()
 
-#Desligar o computador em 1 minuto
-os.system("shutdown /s") 
+def cancel_shutdown():
+    if shutdown == 'sim':
+        os.system("shutdown /a")
 
-#Desligar o computador agora
-os.system("shutdown /s /t 0")
+cancel_shutdown()
 
-#Cancelar o desligamento
-os.system("shutdown /a")
+
+
+# #Desligar o computador em 1 minuto
+# os.system("shutdown /s") 
+
+# #Desligar o computador agora
+# os.system("shutdown /s /t 0")
+
+# #Cancelar o desligamento
+# os.system("shutdown /a")
